public folder css/js and modules
serve-static, & path are npm modules...
I also installed 'serve' but didn't have it as a requirement. It might work without it, but did not test that part. Any file path with '/' so all of them, can read from the folder called 'public'. You have to have the path module in order to use the .join function. From what I read, this is the better way to do it, even if the public file were to move within the main directory. I have not tested that though.  The port change was arbitrarily changed for testing purposes.